cineproject/backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
python-decouple==3.8
